added the ft_substr.c and ft_strlen.c

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: stigkas <stigkas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 16:53:36 by stigkas           #+#    #+#             */
-/*   Updated: 2023/10/31 16:53:49 by stigkas          ###   ########.fr       */
+/*   Updated: 2023/11/01 08:59:25 by stigkas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
